refactor: split control plane connect auth
